mac compilation fix and api to query device availability

+++ icicle_v3/src/curves/montgomery_conversion.cpp
@@ -1,3 +1,4 @@
+#include "icicle/vec_ops.h"
 #include "icicle/curves/montgomery_conversion.h"
 #include "icicle/dispatcher.h"
 #include "icicle/curves/curve_config.h"
@@ -10,14 +11,14 @@ namespace icicle {
   ICICLE_DISPATCHER_INST(AffineConvertMont, affine_convert_montgomery, AffineConvertMontImpl);
 
   extern "C" eIcicleError CONCAT_EXPAND(CURVE, affine_convert_montgomery)(
-    const affine_t* input, size_t n, bool is_into, const VecOpsConfig& config, affine_t* output)
+    const affine_t* input, uint64_t n, bool is_into, const VecOpsConfig& config, affine_t* output)
   {
     return AffineConvertMont::execute(input, n, is_into, config, output);
   }
 
   template <>
   eIcicleError
-  convert_montgomery(const affine_t* input, size_t n, bool is_into, const VecOpsConfig& config, affine_t* output)
+  convert_montgomery(const affine_t* input, uint64_t n, bool is_into, const VecOpsConfig& config, affine_t* output)
   {
     return CONCAT_EXPAND(CURVE, affine_convert_montgomery)(input, n, is_into, config, output);
   }
@@ -44,14 +45,14 @@ namespace icicle {
   ICICLE_DISPATCHER_INST(ProjectiveConvertMont, projective_convert_montgomery, ProjectiveConvertMontImpl);
 
   extern "C" eIcicleError CONCAT_EXPAND(CURVE, projective_convert_montgomery)(
-    const projective_t* input, size_t n, bool is_into, const VecOpsConfig& config, projective_t* output)
+    const projective_t* input, uint64_t n, bool is_into, const VecOpsConfig& config, projective_t* output)
   {
     return ProjectiveConvertMont::execute(input, n, is_into, config, output);
   }
 
   template <>
   eIcicleError convert_montgomery(
-    const projective_t* input, size_t n, bool is_into, const VecOpsConfig& config, projective_t* output)
+    const projective_t* input, uint64_t n, bool is_into, const VecOpsConfig& config, projective_t* output)
   {
     return CONCAT_EXPAND(CURVE, projective_convert_montgomery)(input, n, is_into, config, output);
   }
@@ -60,14 +61,14 @@ namespace icicle {
   ICICLE_DISPATCHER_INST(ProjectiveG2ConvertMont, projective_g2_convert_montgomery, ProjectiveG2ConvertMontImpl);
 
   extern "C" eIcicleError CONCAT_EXPAND(CURVE, projective_g2_convert_montgomery)(
-    const g2_projective_t* input, size_t n, bool is_into, const VecOpsConfig& config, g2_projective_t* output)
+    const g2_projective_t* input, uint64_t n, bool is_into, const VecOpsConfig& config, g2_projective_t* output)
   {
     return ProjectiveG2ConvertMont::execute(input, n, is_into, config, output);
   }
 
   template <>
   eIcicleError convert_montgomery(
-    const g2_projective_t* input, size_t n, bool is_into, const VecOpsConfig& config, g2_projective_t* output)
+    const g2_projective_t* input, uint64_t n, bool is_into, const VecOpsConfig& config, g2_projective_t* output)
   {
     return CONCAT_EXPAND(CURVE, projective_g2_convert_montgomery)(input, n, is_into, config, output);
   }
